requirements.txt
fastapi
pydantic
uvicorn
redis
python-dotenv
passlib
sqlalchemy
psycopg2-binary
fastapi-jwt-auth
python-multipart
cryptography
sqlmodel
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger